fixed a massive bug that sorted the job scores but not the job titles. Also added reasons for the scores
