+++ .gitignore
@@ -7,3 +7,4 @@ google_service_account_credentials.json
 candidate_summary.txt
 dist/
 batch_files
+processed_jobs/
