+++ index.ts
@@ -1,59 +1,17 @@
 import 'dotenv/config';
-import { scrapeJobs } from './services/apifyService.js';
+import { scrapeJobs, parseJobs } from './services/apifyService.js';
 import { readJSONFromFile } from './services/fileService.js';
 import { readJobsFromSheet, writeJobToSheet, writeJobsToSheet } from './services/googleSheetsService.js';
 import { filterDuplicateJobs } from './services/jobUtils.js';
 import { batchProcessJobs } from './services/jobMatchEvaluator.js';
 import { log } from './services/loggingService.js';
 import { evaluateJobMatch } from './services/jobMatchEvaluator.js';
-import { readFileSync, readdirSync } from 'fs';
+import { readFileSync, readdirSync, writeFileSync, existsSync, mkdirSync } from 'fs';
 import { Job, JobAiResponses, UnparsedJobList } from './types.js';
 const candidateSummary = readFileSync('./candidate_summary.txt', 'utf8');
 const useDebugMode = process.env.DEBUG_MODE === "true";
 
 
-function parseJobs(jobLists: UnparsedJobList[]): Job[] {
-    return jobLists.flatMap(jobList => {
-        switch (jobList.actorName) {
-            case 'curious_coder/linkedin-jobs-scraper':
-                return jobList.unparsed_jobs.map((job: any) => ({
-                    title: job.title,
-                    companyName: job.companyName,
-                    location: job.location,
-                    jobUrl: job.link,
-                    pay: job.salaryInfo.join(' - '),
-                    contractType: job.employmentType,
-                    description: job.descriptionText + '\n' + job.companyDescription,
-                    source: `LinkedIn via Apify https://console.apify.com/actors/${jobList.actorId}/information/latest/readme`
-                }));
-            case 'memo23/apify-ziprecruiter-scraper':
-                return jobList.unparsed_jobs.map((job: any) => ({
-                    title: job.Title,
-                    companyName: job.OrgName,
-                    location: job.City,
-                    jobUrl: job.Href,
-                    pay: job.FormattedSalaryShort,
-                    contractType: job.EmploymentType,
-                    description: job.description,
-                    source: `ZipRecruiter via Apify https://console.apify.com/actors/${jobList.actorId}/information/latest/readme`
-                }));
-            case 'curious_coder/indeed-scraper':
-                return jobList.unparsed_jobs.map((job: any) => ({
-                    title: job.displayTitle,
-                    companyName: job.company,
-                    location: job.jobLocationCity,
-                    jobUrl: job.thirdPartyApplyUrl?.replace('indeed.com//', 'indeed.com/'),
-                    pay: job.salarySnippet.text,
-                    contractType: job.jobTypes[0] || '',
-                    description: job.jobDescription,
-                    source: `Indeed via Apify https://console.apify.com/actors/${jobList.actorId}/information/latest/readme`
-                }));
-            default:
-                log('WARN', `No parser found for actor: ${jobList.actorName}`);
-                return [];
-        }
-    });
-}
 async function main() {
     log('INFO', 'Starting Job Scraper Workflow...');
     /* The job scraping is done in parallel. This script will wait for all Apify actors/jobs to finish before the GPT evaluation to prevent GPT from processing duplicate jobs from different searches/Apify actors. 
@@ -102,18 +60,36 @@ async function main() {
              // Step 4: Process all jobs in batch
             const jobResults: JobAiResponses[] = await batchProcessJobs(uniqueJobs, candidateSummary);
 
-            // sort all the jobs by match percentage in descending order
-            jobResults.sort((a, b) => {
-                const scoreA = Number(a.gptJobMatchPercentage || 0);
-                const scoreB = Number(b.gptJobMatchPercentage || 0);
+            // Create paired data to keep jobs and results together
+            const pairedJobData = uniqueJobs.map((job, index) => ({
+                job,
+                result: jobResults[index]
+            }));
+
+            // Sort paired data by match percentage
+            pairedJobData.sort((a, b) => {
+                const scoreA = Number(a.result.gptJobMatchPercentage || 0);
+                const scoreB = Number(b.result.gptJobMatchPercentage || 0);
                 return scoreB - scoreA;
             });
+
+            // Write sorted jobs to debug file
+            const timestamp = new Date().toISOString().replace(/[:.]/g, '-');
+            const debugDir = './processed_jobs';
+            if (!existsSync(debugDir)) {
+                mkdirSync(debugDir);
+            }
+            writeFileSync(
+                `${debugDir}/sorted_jobs_${timestamp}.json`, 
+                JSON.stringify(pairedJobData, null, 2)
+            );
             
             // Step 5: Write to sheets in batches (the "recommended" limit is 2 MB of data per call)
             const BATCH_SIZE = 1000;
-            for (let i = 0; i < uniqueJobs.length; i += BATCH_SIZE) {
-                const jobBatch = uniqueJobs.slice(i, i + BATCH_SIZE);
-                const resultsBatch = jobResults.slice(i, i + BATCH_SIZE);
+            for (let i = 0; i < pairedJobData.length; i += BATCH_SIZE) {
+                const batchPairs = pairedJobData.slice(i, i + BATCH_SIZE);
+                const jobBatch = batchPairs.map(pair => pair.job);
+                const resultsBatch = batchPairs.map(pair => pair.result);
                 
                 await writeJobsToSheet(jobBatch, resultsBatch)
